Data.HasCacBDD

Description

Haskell bindings for CacBDD, a BDD Package with Dynamic Cache Management.

Types

data Bdd

The CacBDD datatype has no structure because from our perspective BDDs are just pointers.

type Assignment = [(Int, Bool)]

An assignment of boolean values to variables/integers.

Creation of new BDDs

top :: Bdd

True constant

bot :: Bdd

False constant

var :: Int -> Bdd

Variable, indexed by any integer from 0 to 1.000.000 FIXME: Segfaults if n is negative or out of range. Can we add a safety check without affecting performance?

Combination and Manipulation of BDDs

neg :: Bdd -> Bdd

Negation

con :: Bdd -> Bdd -> Bdd

Conjunction

dis :: Bdd -> Bdd -> Bdd

Disjunction

imp :: Bdd -> Bdd -> Bdd

Implication, via disjunction and negation. Somehow this is faster than calling LessEqual?

equ :: Bdd -> Bdd -> Bdd

Equivalence aka Biimplication

xor :: Bdd -> Bdd -> Bdd

Exclusive Or

conSet :: [Bdd] -> Bdd

Big Conjunction

disSet :: [Bdd] -> Bdd

Big Disjunction

xorSet :: [Bdd] -> Bdd

Big Xor

exists :: Int -> Bdd -> Bdd

Existential Quantification

forall :: Int -> Bdd -> Bdd

Universal Quantification

forallSet :: [Int] -> Bdd -> Bdd

Big Universal Quantification

existsSet :: [Int] -> Bdd -> Bdd

Big Existential Quantification

restrict :: Bdd -> (Int, Bool) -> Bdd

Restrict a single variable to a given value

restrictSet :: Bdd -> Assignment -> Bdd

Restrict with a (partial) assignment

restrictLaw :: Bdd -> Bdd -> Bdd

Restrict with a law. Note that the law is the second parameter!

ifthenelse :: Bdd -> Bdd -> Bdd -> Bdd

If ... then ... else ...

gfp :: (Bdd -> Bdd) -> Bdd

Greatest fixpoint for a given operator.

relabel :: [(Int, Int)] -> Bdd -> Bdd

Relabel variables according to the given mapping. Note that the mapping list must be sorted!

relabelFun :: (Int -> Int) -> Bdd -> Bdd

Relabel variables according to the given function.

substit :: Int -> Bdd -> Bdd -> Bdd

Substitute a BDD for a given variable in another BDD.

substitSimul :: [(Int, Bdd)] -> Bdd -> Bdd

Simultaneous substitution of BDDs for variables. Note that this is not the same as folding substit.

Evaluation

evaluate :: Bdd -> Assignment -> Maybe Bool

Evaluate a BDD given an assignment. Returns Nothing if the assignment does not cover allVarsOf b.

evaluateFun :: Bdd -> (Int -> Bool) -> Bool

Evaluate a BDD given a total assignment function.

Get satisfying assignments

allSats :: Bdd -> [Assignment]

Get all satisfying assignments. These will be partial, i.e. only contain (a subset of) the variables that actually occur in the BDD.

allSatsWith :: [Int] -> Bdd -> [Assignment]

Get all complete assignments, given a set of all variables. In particular this will include variables not in the BDD.

satCountWith :: [Int] -> Bdd -> Int

Given a set of all variables, get the number of satisfying assignments. Note that allvars must be nub'd and sorted.

anySat :: Bdd -> Maybe Assignment

Get the lexicographically smallest satisfying assignment, if there is any.

anySatWith :: [Int] -> Bdd -> Maybe Assignment

Given a set of all variables, get the lexicographically smallest complete satisfying assignment, if there is any.

Variables

firstVarOf :: Bdd -> Maybe Int

The first variable of a given BDD, if there is one.

maxVarOf :: Bdd -> Maybe Int

The maximum variable of a given BDD, if there is one.

allVarsOf :: Bdd -> [Int]

All variables in a given BDD, *not* sorted, lazy.

allVarsOfSorted :: Bdd -> [Int]

All variables in a given BDD, sorted, *not* lazy.

Sub-BDDs and length

thenOf :: Bdd -> Bdd

Then-branch of a given BDD, setting firstVarOf to True.

elseOf :: Bdd -> Bdd

Else-branch of a given BDD, setting firstVarOf to False.

subsOf :: Bdd -> [Bdd]

List all the sub-BDDs of a given BDD.

sizeOf :: Bdd -> Int

Size of the BDD, should be the number of non-terminal nodes.

Show and convert to trees

data BddTree

A simple tree definition to show BDDs as text.

Constructors

Bot

Top

Var Int BddTree BddTree

unravel :: Bdd -> BddTree

Convert a BDD to a tree.

ravel :: BddTree -> Bdd

Convert a tree to a BDD.

Print some debugging information

maximumvar :: Int

The maximum number of variables

showInfo :: IO ()

Show internal statistics.